Define String#getord(offset) for accessing a single byte at a String offset
Strings have changed significantly in Ruby 1.9.  The main issues related
to Grit are encodings, and the [] method.  The GitRuby implementation
currently uses code such as "buf[offset]" to access a byte (1 uint8_t byte in C)
of data. It would be nice to just use String#getbyte(offset), but this method is not
sensitive to the string encoding.  The default encoding for Ruby strings
is UTF-8, and I can't figure out how to make sure that all of the encodings
are correct such that String#getbyte will return the right value.

In Ruby 1.8, String[offset] returns the integer code of the character at offset.
In Ruby 1.9, String[offset] returns the actual character at offset. To get the integer
value, we need to call String#ord, which returns the int code of the first char.
Thus, to get Ruby 1.8 behavior in Ruby 1.9, we would need to call String[offset].ord.
My solution for now is just to define a getord(offset) method in a version dependent
manner, and use this in place of String[offset].  It's somewhat hackish, but
I can't think of a better solution right now.