Pass correct --work-tree to git
In order to run Git against a work tree other than the current working
directory, you need to pass the --work-tree parameter in.  Previous
versions of Grit::Git tried doing this by simply srtipping off a path
component matching /.git from the git_dir.  This is clearly wrong, as
it means that bare repos would be considered their own working dirs,
which I'm sure will lead to all kinds of problems. To fix this, the
--work-tree parameter was simply removed entirely from Grit::Git,
which is also wrong, as it means that when you're using a non-bare
repo, Grit::Git will just write all over your current working directory
instead of the working tree for that repo.

This patch updates Grit::Git to accept a work tree parameter, and
passes the correct working directory from Grit::Repo to Grit::Git when
working with a non-bare repo.  For bare repos, no working directory is
passed in, and so no --work-tree is passed to Git.