Fix Actor time zone calculation for half-hour time zones
The Actor unit tests had been using a hard-coded time zone compared
against the current local time, which meant they would fail if run in
a different time zone.  Since the Time object doesn't let you
construct times in a zone other than local or UTC, I've changed the
test to use a stub object in a fixed time zone, instead of the current
local time.

This revealed a problem in the handling of fractional time zones; the
remainder when converting from seconds to hours was used directly as
minutes, instead of being converted from seconds to minutes.  This
patch converts the remainder into minutes.