fix relative alternates paths
Relative paths in a <repo>.git/objects/info/alternates file are to
be expanded relative to the <repo>.git/objects directory.
Previously, they were expanded relative to the <repo>.git directory.

The current code checks in both locations for backward compatibility
with grit <= 2.3.0 but I'd rather rip it out completely.