Fixed bug with improper shifting in renamed diffs
* Changed the number 2 to 3 when shifting the lines in a copied or
renamed diff. Made sure to add assertions to the diff tests to
check for '-' as the first character for every non-empty diff.
* Added a condition to check for empty diffs and skip if necessary.

Until now, a non 100% renamed file gives it's diff starting with the
line 'index ....' as opposed to '--- a/...' in case of other diffs.
This commit solved that issue.